analyze: GitLab config modal positioning issue
I have a ui issue that I need you to analyse. The modal to set up the gitlab config is broken. it should be in the middle of the screen in a defined container space. Currently, it is at the top of the screen, barely visible (like it has been pushed to the back) but scrollable. Analyse the codebase thoroughly to pinpoint the exact issue that is causing this problem. Do not make any changes to the codebase for now.

```git-revs
60f9bdd  (Base revision)
bfa1401  Remove GitLabSettings modal from UserProfile component to eliminate positioning context constraints
371029c  Add GitLabSettings import to dashboard page
8ecfe4c  Add GitLabSettings modal to system health warning section
c8007ad  Add GitLabSettings modal to dashboard with projects section
HEAD     Add GitLabSettings modal to empty state section
```

codemcp-id: 13-analyze-gitlab-config-modal-positioning-issue